allowing student to continue without last LL spec
The last LL spec, `forEach`, can be a challenge. If a student is taking too long they never attempt `Hash`, even though `Hash` can be implemented without an LL `forEach` (it helps on the last spec of Hash but isn't strictly necessary). Edited the comments to make this more explicit.